Reorganization/commenting code
Code has been reorganized and is commented, still for 2 window setup.
**THIS VERSION DOES NOT WORK.  Since I will be updating the entire setup
to to be 1 window I decided to commit this version and then start
working on that version

Next steps:
Going to reorganize everything to make it work in a 1 window display.
Will require updating the html and additional reorganization of the
code.